+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.7.1
+
+- Using `/characters/affiliation/` and `/universe/names/` instead of `/characters/{character_id}/` 
+  to update characters (much shorter cache time and faster)
+
 ## 1.7.0
 
 21 Nov 2019